Cut the costs that scale with the number of partition entities
Four changes, all aimed at the large-partition-count regime, where the
entity count reaches ~800k on a 4.6M-element mesh at 32768 partitions:

- divideNonConnectedEntities() called Graph::eraseVertex() once per partition
  entity, and _vertex is sized by the largest node number in the mesh, so
  resetting it was O(#entities x #nodes). Record the entries actually used
  and reset only those.

- The per-entity graphs are reused from one entity to the next, so they only
  need to be as large as the largest single entity. They were sized from the
  model, which made each dimension allocate and zero an eind array covering
  the whole mesh (~150 MB three times over on the 4.6M-element mesh).

- GModel::remove() scanned the entity container linearly although it is
  sorted by tag, so every removal was O(#entities). Look the entity up
  instead, falling back to a scan in case a tag was changed after insertion.

- GModel::getMaxElementaryNumber() materialized a vector of every entity in
  the model on each call. Scan the relevant containers directly.

Also widen MElement::_partition from short to int: setPartition() silently
overflowed above 32767 partitions, which was already wrong for MSH2 output
and ConvertOldPartitioningToNewOne. It is free, as those bytes were padding.

Partitioning / topology phases, wall, against master:

  multivol 1.96M   2048 parts   7.5/ 4.8s -> 4.9/3.1s   1.5x
                   8192 parts  12.8/11.5s -> 10.2/6.0s  1.5x
  cube     4.58M   2048 parts  37.5/ 9.7s -> 13.7/7.0s  2.3x
                   8192 parts  43.2/22.8s -> 18.6/11.4s 2.2x

Of which this commit roughly halves the topology phase at high partition
counts (cube at 8192: 22.5s -> 11.7s on top of the dual-graph work).

Output is bit-identical to the previous commit across 25 configurations plus
the no-topology path, and equal as a multiset on the ghost path.

+++ src/geo/MElement.h
@@ -33,8 +33,10 @@ private:
   // to change once a mesh has been generated, unless the mesh is explicitly
   // renumbered)
   std::size_t _num;
-  // the number of the mesh partition the element belongs to
-  short _partition;
+  // the number of the mesh partition the element belongs to (this used to be a
+  // short, which silently overflowed above 32767 partitions; widening it is
+  // free, as the bytes were padding)
+  int _partition;
   // a visibility flag
   char _visible;
 
@@ -88,7 +90,7 @@ public:
 
   // get/set the partition to which the element belongs
   virtual int getPartition() const { return _partition; }
-  virtual void setPartition(int num) { _partition = (short)num; }
+  virtual void setPartition(int num) { _partition = num; }
 
   // get/set the visibility flag
   virtual char getVisibility() const;
